+++ src/components/panel/annotations/filters/FilterTree.tsx
@@ -12,19 +12,27 @@ const FilterTree: FC<Props> = ({ nodes, onToggle }) => {
   const { usePanelTranslation } = usePanel()
   const { t } = usePanelTranslation()
 
+  const anyChildren = nodes.some((node) => node.items && node.items.length > 0)
+
   if (nodes.length === 0) {
     return <div className="h-16 px-3 text-sm text-muted-foreground flex justify-center items-center">{ t('no_annotation_filters') }</div>
   }
 
-  return nodes.map((node, index) => (
-    <FilterTreeNode
-      key={index}
-      node={node}
-      path={[index]}
-      onToggle={onToggle}
-      depth={0}
-    />
-  ))
+  return <div className="-ml-6">
+    {
+      nodes.map((node, index) => (
+        <FilterTreeNode
+          key={index}
+          node={node}
+          path={[index]}
+          onToggle={onToggle}
+          indented={anyChildren && !(node.items && node.items.length > 0)} // only intend if some sibling has children but this node does not
+        />
+      ))
+    }
+  </div>
+
+
 }
 
 export default FilterTree