+++ src/components/ui/badge.tsx
@@ -3,6 +3,7 @@ import { Slot } from '@radix-ui/react-slot'
 import { cva, type VariantProps } from 'class-variance-authority'
 
 import { cn } from '@/lib/utils'
+import { memo } from 'react'
 
 const badgeVariants = cva(
   'inline-flex items-center justify-center rounded-md border px-1 py-0.5 text-xs font-medium w-fit whitespace-nowrap shrink-0 [&>svg]:size-3 gap-1 [&>svg]:pointer-events-none focus-visible:border-ring focus-visible:ring-ring/50 focus-visible:ring-[3px] aria-invalid:ring-destructive/20 dark:aria-invalid:ring-destructive/40 aria-invalid:border-destructive transition-[color,box-shadow] overflow-hidden',
@@ -11,6 +12,8 @@ const badgeVariants = cva(
       variant: {
         default:
           'border-transparent bg-primary text-primary-foreground [a&]:hover:bg-primary/90',
+        accent:
+          'border-transparent bg-accent text-accent-foreground [a&]:hover:bg-muted',
         secondary:
           'border-transparent bg-secondary text-secondary-foreground [a&]:hover:bg-secondary/90',
         destructive:
@@ -25,7 +28,7 @@ const badgeVariants = cva(
   }
 )
 
-const Badge = ({
+const Badge = memo(({
   className,
   variant,
   asChild = false,
@@ -41,6 +44,6 @@ const Badge = ({
       {...props}
     />
   )
-}
+})
 
 export { Badge, badgeVariants }