+++ src/utils/config/default-config.ts
@@ -3,7 +3,6 @@ import { TidoConfig } from '@/types'
 const defaultConfig: TidoConfig = {
   allowNewCollections: true,
   container: '#app',
-  defaultPanelMode: 'swap',
   lang: 'en',
   panels: [],
   rootCollections: [],
@@ -16,9 +15,18 @@ const defaultConfig: TidoConfig = {
   },
   title: '',
   translations: {},
-  panelModes: ['swap', 'split', 'text', 'image'],
-  defaultAnnotationsMode: 'aligned',
-  annotations: {}
+  panelViews: [
+    {
+      label: 'Image',
+      view: 'image'
+    }, {
+      label: 'Text',
+      view: 'text'
+    }
+  ],
+  annotations: {
+    defaultMode: 'aligned'
+  }
 }
 
 export {